+++ skoolkit/snapshot.py
@@ -1,4 +1,4 @@
-# Copyright 2009-2013, 2015-2019 Richard Dymond (rjdymond@gmail.com)
+# Copyright 2009-2013, 2015-2020 Richard Dymond (rjdymond@gmail.com)
 #
 # This file is part of SkoolKit.
 #
@@ -83,7 +83,7 @@ def get_snapshot(fname, page=None):
     mem.extend(ram)
     return mem
 
-def make_snapshot(fname, org=None, start=16384, end=65536, page=None):
+def make_snapshot(fname, org, start=0, end=65536, page=None):
     snapshot_reader = get_snapshot_reader()
     if snapshot_reader.can_read(fname):
         return snapshot_reader.get_snapshot(fname, page), max(16384, start), end